requirements.txt
flask==3.0.3
google-generativeai==0.8.3
torch==2.8.0
requests==2.32.4
pandas==2.2.2
scikit-learn==1.5.2